# improve matrix_shift_add