In the case of multiples of 8, I did simple software pipelining optimization.
There is no effect in other cases.