@lm-konda 2018-11-27
performance is below.

before
```
TotalRunTime,292640,  sum:292.64ms
kn2row,156757,  sum:156.757ms
kn2row-1x1,2098,  sum:2.098ms
kn2row-buf,4,  sum:0.004ms
matrix_multiplication,98908,2089,  sum:100.997ms
matrix_shift_add_f1,3558,  sum:3.558ms
matrix_shift_add_f2,54242,  sum:54.242ms
```
after
```
TotalRunTime,286993,  sum:286.993ms
kn2row,150982,  sum:150.982ms
kn2row-1x1,2098,  sum:2.098ms
kn2row-buf,4,  sum:0.004ms
matrix_multiplication,99020,2089,  sum:101.109ms
matrix_shift_add_f1,3559,  sum:3.559ms
matrix_shift_add_f2,48356,  sum:48.356ms
```